#6 Render namespace template in custom K8s resources
A dogu's custom K8s resource may contain go template strings that should render
the current namespace. This commit adds this render functionality on per-yaml-
document basis. Each document is tried to render the namespace. This new
document in turn should be used for _all_ subsequent actions because it was
substantially modified (esp. for ServiceAccounts which will be returned to the
manager for further processing).
